add some app-web fixes
