+++ services/app-web/package.json
@@ -93,7 +93,7 @@
         "@opentelemetry/resources": "^1.18.1",
         "@opentelemetry/sdk-trace-web": "^1.18.1",
         "@opentelemetry/semantic-conventions": "^1.18.1",
-        "@tanstack/react-table": "^8.8.5",
+        "@tanstack/react-table": "^8.19.3",
         "@trussworks/react-uswds": "^4.0.0",
         "@types/path-browserify": "^1.0.0",
         "@vitejs/plugin-react": "^4.3.0",
@@ -112,6 +112,7 @@
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-error-boundary": "^4.0.10",
+        "react-router": "^6.23.1",
         "react-router-dom": "^6.23.1",
         "react-select": "^5.7.7",
         "sass": "^1.49.11",
@@ -144,8 +145,11 @@
         "@testing-library/user-event": "^14.4.3",
         "@types/classnames": "^2.2.11",
         "@types/jest": "^29.5.6",
+        "@types/node": "^20.14.12",
+        "@types/qs": "^6.9.15",
         "@types/react": "^18.0.14",
         "@types/react-dom": "^18.0.5",
+        "@types/react-router": "^5.1.20",
         "@types/react-test-renderer": "^18.0.0",
         "@types/url-parse": "^1.4.3",
         "@types/uuid": "^9.0.0",
