+++ services/app-web/src/components/HealthPlanPackageTable/HealthPlanPackageTable.tsx
@@ -6,11 +6,9 @@ import {
     getCoreRowModel,
     getFilteredRowModel,
     getSortedRowModel,
-    RowData,
     useReactTable,
     getFacetedUniqueValues,
     Column,
-    FilterFn,
 } from '@tanstack/react-table'
 import { useAtom } from 'jotai/react'
 import { atomWithHash } from 'jotai-location'
@@ -30,16 +28,6 @@ import { InfoTag, TagProps } from '../InfoTag/InfoTag'
 import { pluralize } from '../../common-code/formatters'
 import { DoubleColumnGrid } from '../DoubleColumnGrid'
 import { NavLinkWithLogging } from '../TealiumLogging/Link'
-
-declare module '@tanstack/table-core' {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    interface ColumnMeta<TData extends RowData, TValue> {
-        dataTestID: string
-    }
-    interface FilterFns {
-        dateRangeFilter: FilterFn<unknown>
-    }
-}
 
 export type RateInDashboardType = {
     id: string
