+++ services/app-web/tsconfig.json
@@ -2,7 +2,13 @@
     "extends": "../../tsconfig.json",
     "compilerOptions": {
         "target": "ESNext",
-        "types": ["cypress", "node", "vite/client", "vitest/globals"],
+        "types": [
+            "cypress",
+            "node",
+            "vite/client",
+            "vitest/globals",
+            "@tanstack/react-table"
+        ],
         "lib": ["es6", "dom", "dom.iterable", "esnext"],
         "allowJs": true,
         "skipLibCheck": true,
@@ -14,5 +20,5 @@
         "noEmit": true,
         "jsx": "react-jsx"
     },
-    "include": ["src", "vite.config.ts", "vite-env.d.ts"]
+    "include": ["src/**/*", "vite.config.ts", "vite-env.d.ts"]
 }
